[refactor] main/features/simpleInformation 측 파일경로 재지정
