--- src/mainPage/features/simpleInformation/contentList.json
@@ -1,56 +1,0 @@
-{
-  "content": [
-    {
-      "src": "carimg1.png",
-      "title": "독창적인 디자인",
-      "desc": [
-        "아이오닉 브랜드를 상징하는\n",
-        "파라메트릭 픽셀 램프 디자인",
-        "으로\n유니크한 이미지를 구현합니다."
-      ],
-      "sub": "프로젝션 타입과 MFR 타입 중 선택 가능"
-    },
-    {
-      "src": "carimg2.png",
-      "title": "전용 전기차 플랫폼(E-GMP)",
-      "desc": [
-        "",
-        "새로워진 전기차 플랫폼 E-GMP",
-        "는\n알루미늄 압출재를 이용해\n구조적 안정성을 높였습니다."
-      ],
-      "sub": "연출된 이미지로 실제 작동 사양과 다를 수 있음"
-    },
-    {
-      "src": "carimg3.png",
-      "title": "인터랙티브 픽셀 라이트",
-      "desc": [
-        "",
-        "신규 디자인의 스티어링 휠",
-        "로\n아이오닉5만의 차별화된\n주행 경험을 제공합니다."
-      ],
-      "sub": "시동 / 충전 중 / 후진 중 표시 가능"
-    },
-    {
-      "src": "carimg4.png",
-      "title": "증강현실 내비게이션",
-      "desc": [
-        "주행에 필요한 각종 정보들을\n",
-        "증강현실 기술",
-        "을 통해\n직관적으로 제공합니다."
-      ],
-      "sub": "인포테인먼트 시스템 화면 이미지는 업데이트에 따라 변동 가능"
-    },
-    {
-      "src": "carimg5.png",
-      "title": "디지털 사이드 미러",
-      "desc": [
-        "보다 슬림해진 ",
-        "디지털 사이드 미러",
-        "는\n카메라와 ",
-        "OLED 모니터",
-        "를 통해\n선명한 후방 시야를 제공합니다."
-      ],
-      "sub": "모니터는 내부 운전석에 위치"
-    }
-  ]
-}
