+++ src/mainPage/shared/makeHighlight.jsx
