+++ src/mainPage/shared/scroll/constants.js
@@ -1,4 +1,5 @@
 // scroll section constants
+export const OTHER_SECTION = 0;
 export const INTERACTION_SECTION = 1;
 export const DETAIL_SECTION = 2;
 export const COMMENT_SECTION = 3;
